move to analyses.js
